Bug - systemd::unit::ensure absent extends ill behaved
If a systemd unit is extended and ensure=>absent, this definition should
not be stopping/disabling the unit but rather simply removing the
extension.  It would probably be good to refactor this to provide
a Define[systemd::unit::extension] but that'll have to wait.

+++ manifests/unit.pp
@@ -127,14 +127,16 @@ define systemd::unit (
 
     } elsif $ensure == 'absent' {
 
-        exec { "systemctl disable ${target}":
-            before => File[$fqfn],
-            onlyif => "test -e '${fqfn}'",
-        }
+        if $extends == undef {
+            exec { "systemctl disable ${target}":
+                before => File[$fqfn],
+                onlyif => "test -e '${fqfn}'",
+            }
 
-        exec { "systemctl stop ${target}":
-            before => File[$fqfn],
-            onlyif => "test -e '${fqfn}'",
+            exec { "systemctl stop ${target}":
+                before => File[$fqfn],
+                onlyif => "test -e '${fqfn}'",
+            }
         }
 
     } else {
